Correctly use of `std::numeric_limits<T>::epsilon`
Previously, `T` == `double`, which did not match expected behavior from
the original implementation, where `T` == `float`.

+++ trac_ik_lib/src/kdl_tl.cpp
@@ -159,7 +159,7 @@ int ChainIkSolverPos_TL::CartToJnt(const KDL::JntArray &q_init, const KDL::Frame
 
     Subtract(q_out, q_curr, q_out);
 
-    if (q_out.data.isZero(std::numeric_limits<double>::epsilon()))
+    if (q_out.data.isZero(std::numeric_limits<float>::epsilon()))
     {
       if (rr)
       {
